Add option for CPU to iterate sequentially
By default grid_stride_range on the CPU uses a stride equal to
the number of threads, which can be less efficient due to memory
cache misses.  This allows the option (chosen based on preprocessor
definitions) to use a stride of one, and have each CPU use handle
a different block of indices.